[Feat] 팜어스 tab bar 스탭 구현 (#156)
