+++ lib/common/farmus_picture_fix.dart
@@ -4,10 +4,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'theme/farmus_theme_color.dart';
 
 class FarmusPictureFix extends ConsumerWidget {
-  const FarmusPictureFix({super.key, required this.size, this.child});
+  const FarmusPictureFix({super.key, required this.size, this.asset});
 
   final double size;
-  final Widget? child;
+  final String? asset;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -15,12 +15,17 @@ class FarmusPictureFix extends ConsumerWidget {
       width: size,
       height: size,
       decoration: ShapeDecoration(
+        image: asset != null
+            ? DecorationImage(
+                image: AssetImage(asset!),
+                fit: BoxFit.cover,
+              )
+            : null,
         color: FarmusThemeColor.background,
         shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
+          borderRadius: BorderRadius.circular(16.0),
         ),
       ),
-      child: child,
     );
   }
 }
