+++ lib/common/tab_bar/farmclub_tab_bar.dart
@@ -1,5 +1,8 @@
 import 'package:farmus/common/tab_bar/primary_tab_bar.dart';
+import 'package:farmus/view/farmclub/component/farmclub_step.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
 class FarmclubTabBar extends StatelessWidget {
   const FarmclubTabBar({
@@ -10,7 +13,54 @@ class FarmclubTabBar extends StatelessWidget {
   Widget build(BuildContext context) {
     return const PrimaryTabBar(
       tab: ['현재', '이전', '다음'],
-      tabView: [Text('data'), Text('data'), Text('data')],
+      tabView: [
+        SingleChildScrollView(
+          child: Column(
+            children: [
+              Padding(
+                padding: EdgeInsets.all(16.0),
+                child: FarmclubStep(
+                  stepImages: [
+                    'assets/image/img_home_motivation.png',
+                    'assets/image/img_home_motivation.png'
+                  ],
+                ),
+              ),
+              Text('data')
+            ],
+          ),
+        ),
+        SingleChildScrollView(
+          child: Column(
+            children: [
+              Padding(
+                padding: EdgeInsets.all(16.0),
+                child: FarmclubStep(
+                  stepImages: [
+                    'assets/image/img_home_motivation.png',
+                    'assets/image/img_home_motivation.png'
+                  ],
+                ),
+              ),
+              Padding(
+                padding: EdgeInsets.all(16.0),
+                child: FarmclubStep(
+                  stepImages: [],
+                ),
+              ),
+              Padding(
+                padding: EdgeInsets.all(16.0),
+                child: FarmclubStep(
+                  stepImages: [],
+                ),
+              ),
+            ],
+          ),
+        ),
+        FarmclubStep(
+          stepImages: [],
+        )
+      ],
     );
   }
 }
